requirements.txt
aiogram==3.15.0
python-dotenv==1.0.1
aiohttp==3.11.11
aiofiles==24.1.0
